Simple script to test how you can effectively save plots of the predictors to a list within a for loop
Also tested if you could use one of the apply functions. However, basic R plotting routines do not work with loops/apply, since plots are directly sent to the graphics device and can not be stored.
With ggplot, you need to place the ith column somewhere else than in the aesthetics, otherwise only the last plot is saved.