Finished rewriting the whole script.
Made script more sophisticated (o.a. added a function to plot scree plots of the explained variance of PC's).

Found strange results for the PC's of the test set. See scree and correlation plots. Can not idenentify where the problem lies. Maybe there are issues because a 6h + 1.5 subset of data is too small?